Added delete taks functionality
