+++ frontend/src/pages/Dashboard.js
@@ -1,17 +1,28 @@
-import React, { useEffect, useState } from 'react'
+import React, { useMemo, useState } from 'react'
 import TaskModal from '../components/TaskModal'
 import TaskView from '../components/TaskView'
+import apiClient from '../services/api'
 import { useAuth } from '../services/AuthProvider'
 
 const Dashboard = () => {
     const { loggedIn, loggedInUser, logout } = useAuth()
+    const [tasks, setTasks] = useState([])
     const [showModal, setShowModal] = useState(false)
     const [editId, setEditId] = useState(0)
     const [editState, setEditState] = useState(false)
-    
-    useEffect(() => {
-        console.log(loggedIn)
-    }, [])
+
+    const data = useMemo(() => {
+        apiClient({
+            method: 'get',
+            url: `/api/v1/tasks/${loggedInUser.id}`
+        }).then(response => {
+            setTasks(response.data)
+        }).catch(error => {
+            console.log(error)
+        })
+
+        setEditState(false)
+    }, [editState])
 
     return (
         <div>
@@ -20,7 +31,7 @@ const Dashboard = () => {
                 <button type='button' onClick={logout}>Logout</button>
             </div>
             <hr />
-            <TaskView userId={loggedInUser.id} editState={editState} setEditId={setEditId} setShowModal={setShowModal} />
+            <TaskView tasks={tasks} editState={editState} setEditState={setEditState} setEditId={setEditId} setShowModal={setShowModal} />
             <TaskModal showModal={showModal} setShowModal={setShowModal} editId={editId} setEditState={setEditState} />
         </div>
     )
